split compose deploy from deploy.go
